+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 
 ## Unreleased
+### Added
+- Organization social_media field
+- Facility social_media field
+- InternetExchange social_media field
+- Network social_media field
+- Campus social_media field
+- Carrier social_media field
+### Removed
+- django2.2 support
+- django3.0 support
 
 
 ## 2.16.0